Add files via upload
In today's world, Driver drowsiness is a persistent issue and is the cause of 10% of all car crashes. In order to solve this problem, I developed Eye Spy, and app that alerts sleepy drivers to wake up by sounding an alarm. The implemntation of this app has the potential so save lives caused by driver fatigue.